refactor: unify int type

+++ internal/database/online/test/util.go
@@ -48,9 +48,9 @@ func init() {
 			Revision: &types.Revision{ID: 3, GroupID: 1},
 			Entity:   &types.Entity{ID: 5, Name: "user", Length: 4},
 			Data: []types.RawFeatureValueRecord{
-				newRecord([]interface{}{"3215", int16(18), "F"}),
-				newRecord([]interface{}{"3216", int16(29), nil}),
-				newRecord([]interface{}{"3217", int16(44), "M"}),
+				newRecord([]interface{}{"3215", int64(18), "F"}),
+				newRecord([]interface{}{"3216", int64(29), nil}),
+				newRecord([]interface{}{"3217", int64(44), "M"}),
 			},
 		}
